sql/catalog/resolver: extract resolver logic into subpackage
Release note: None
