+++ pkg/sql/alter_sequence.go
@@ -14,6 +14,7 @@ import (
 	"context"
 
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/resolver"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
@@ -28,7 +29,7 @@ type alterSequenceNode struct {
 // AlterSequence transforms a tree.AlterSequence into a plan node.
 func (p *planner) AlterSequence(ctx context.Context, n *tree.AlterSequence) (planNode, error) {
 	seqDesc, err := p.ResolveMutableTableDescriptorEx(
-		ctx, n.Name, !n.IfExists, ResolveRequireSequenceDesc,
+		ctx, n.Name, !n.IfExists, resolver.ResolveRequireSequenceDesc,
 	)
 	if err != nil {
 		return nil, err
