+++ pkg/sql/resolver.go
@@ -16,45 +16,21 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/resolver"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/cat"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/schema"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/errors"
 )
 
-// SchemaResolver abstracts the interfaces needed from the logical
-// planner to perform name resolution below.
-//
-// We use an interface instead of passing *planner directly to make
-// the resolution methods able to work even when we evolve the code to
-// use a different plan builder.
-// TODO(rytaft,andyk): study and reuse this.
-type SchemaResolver interface {
-	tree.ObjectNameExistingResolver
-	tree.ObjectNameTargetResolver
-
-	Txn() *kv.Txn
-	LogicalSchemaAccessor() catalog.Accessor
-	CurrentDatabase() string
-	CurrentSearchPath() sessiondata.SearchPath
-	CommonLookupFlags(required bool) tree.CommonLookupFlags
-	ObjectLookupFlags(required bool, requireMutable bool) tree.ObjectLookupFlags
-	LookupTableByID(ctx context.Context, id sqlbase.ID) (catalog.TableEntry, error)
-}
-
-var _ SchemaResolver = &planner{}
-
-var errNoPrimaryKey = pgerror.Newf(pgcode.NoPrimaryKey,
-	"requested table does not have a primary key")
+var _ resolver.SchemaResolver = &planner{}
 
 // ResolveUncachedDatabaseByName looks up a database name from the store.
 func (p *planner) ResolveUncachedDatabaseByName(
@@ -68,77 +44,94 @@ func (p *planner) ResolveUncachedDatabaseByName(
 	return res, err
 }
 
-// GetObjectNames retrieves the names of all objects in the target database/
-// schema. If explicitPrefix is set, the returned table names will have an
-// explicit schema and catalog name.
-func GetObjectNames(
-	ctx context.Context,
-	txn *kv.Txn,
-	sc SchemaResolver,
-	codec keys.SQLCodec,
-	dbDesc *DatabaseDescriptor,
-	scName string,
-	explicitPrefix bool,
-) (res TableNames, err error) {
-	return sc.LogicalSchemaAccessor().GetObjectNames(ctx, txn, codec, dbDesc, scName,
-		tree.DatabaseListFlags{
-			CommonLookupFlags: sc.CommonLookupFlags(true /* required */),
-			ExplicitPrefix:    explicitPrefix,
-		})
+// runWithOptions sets the provided resolution flags for the
+// duration of the call of the passed argument fn.
+//
+// This is meant to be used like this (for example):
+//
+// var someVar T
+// var err error
+// p.runWithOptions(resolveFlags{skipCache: true}, func() {
+//    someVar, err = ResolveExistingTableObject(ctx, p, ...)
+// })
+// if err != nil { ... }
+// use(someVar)
+func (p *planner) runWithOptions(flags resolveFlags, fn func()) {
+	if flags.skipCache {
+		defer func(prev bool) { p.avoidCachedDescriptors = prev }(p.avoidCachedDescriptors)
+		p.avoidCachedDescriptors = true
+	}
+	fn()
 }
 
-// ResolveExistingTableObject looks up an existing object.
-// If required is true, an error is returned if the object does not exist.
-// Optionally, if a desired descriptor type is specified, that type is checked.
-//
-// The object name is modified in-place with the result of the name
-// resolution, if successful. It is not modified in case of error or
-// if no object is found.
-func ResolveExistingTableObject(
-	ctx context.Context,
-	sc SchemaResolver,
-	tn *TableName,
-	lookupFlags tree.ObjectLookupFlags,
-	requiredType ResolveRequiredType,
-) (res *ImmutableTableDescriptor, err error) {
-	// TODO: As part of work for #34240, an UnresolvedObjectName should be
-	//  passed as an argument to this function.
-	un := tn.ToUnresolvedObjectName()
-	desc, prefix, err := resolveExistingObjectImpl(ctx, sc, un, lookupFlags, requiredType)
-	if err != nil || desc == nil {
-		return nil, err
-	}
-	tn.ObjectNamePrefix = prefix
-	return desc.(*ImmutableTableDescriptor), nil
+type resolveFlags struct {
+	skipCache bool
 }
 
-// ResolveMutableExistingTableObject looks up an existing mutable object.
-// If required is true, an error is returned if the object does not exist.
-// Optionally, if a desired descriptor type is specified, that type is checked.
-//
-// The object name is modified in-place with the result of the name
-// resolution, if successful. It is not modified in case of error or
-// if no object is found.
-func ResolveMutableExistingTableObject(
-	ctx context.Context,
-	sc SchemaResolver,
-	tn *TableName,
-	required bool,
-	requiredType ResolveRequiredType,
-) (res *MutableTableDescriptor, err error) {
-	lookupFlags := tree.ObjectLookupFlags{
-		CommonLookupFlags: tree.CommonLookupFlags{Required: required},
-		RequireMutable:    true,
+func (p *planner) ResolveMutableTableDescriptor(
+	ctx context.Context, tn *TableName, required bool, requiredType resolver.ResolveRequiredType,
+) (table *MutableTableDescriptor, err error) {
+	return resolver.ResolveMutableExistingTableObject(ctx, p, tn, required, requiredType)
+}
+
+func (p *planner) ResolveUncachedTableDescriptor(
+	ctx context.Context, tn *TableName, required bool, requiredType resolver.ResolveRequiredType,
+) (table *ImmutableTableDescriptor, err error) {
+	p.runWithOptions(resolveFlags{skipCache: true}, func() {
+		lookupFlags := tree.ObjectLookupFlags{CommonLookupFlags: tree.CommonLookupFlags{Required: required}}
+		table, err = resolver.ResolveExistingTableObject(ctx, p, tn, lookupFlags, requiredType)
+	})
+	return table, err
+}
+
+func (p *planner) ResolveUncachedDatabase(
+	ctx context.Context, un *tree.UnresolvedObjectName,
+) (res *UncachedDatabaseDescriptor, namePrefix tree.ObjectNamePrefix, err error) {
+	p.runWithOptions(resolveFlags{skipCache: true}, func() {
+		res, namePrefix, err = resolver.ResolveTargetObject(ctx, p, un)
+	})
+	return res, namePrefix, err
+}
+
+// LookupSchema implements the tree.ObjectNameTargetResolver interface.
+func (p *planner) LookupSchema(
+	ctx context.Context, dbName, scName string,
+) (found bool, scMeta tree.SchemaMeta, err error) {
+	sc := p.LogicalSchemaAccessor()
+	dbDesc, err := sc.GetDatabaseDesc(ctx, p.txn, p.ExecCfg().Codec, dbName, p.CommonLookupFlags(false /*required*/))
+	if err != nil || dbDesc == nil {
+		return false, nil, err
 	}
-	// TODO: As part of work for #34240, an UnresolvedObjectName should be
-	//  passed as an argument to this function.
-	un := tn.ToUnresolvedObjectName()
-	desc, prefix, err := resolveExistingObjectImpl(ctx, sc, un, lookupFlags, requiredType)
-	if err != nil || desc == nil {
-		return nil, err
+	found, _, err = sc.IsValidSchema(ctx, p.txn, p.ExecCfg().Codec, dbDesc.ID, scName)
+	if err != nil {
+		return false, nil, err
 	}
-	tn.ObjectNamePrefix = prefix
-	return desc.(*MutableTableDescriptor), nil
+	return found, dbDesc, nil
+}
+
+// LookupObject implements the tree.ObjectNameExistingResolver interface.
+func (p *planner) LookupObject(
+	ctx context.Context, lookupFlags tree.ObjectLookupFlags, dbName, scName, tbName string,
+) (found bool, objMeta tree.NameResolutionResult, err error) {
+	sc := p.LogicalSchemaAccessor()
+	lookupFlags.CommonLookupFlags = p.CommonLookupFlags(false /* required */)
+	objDesc, err := sc.GetObjectDesc(ctx, p.txn, p.ExecCfg().Settings, p.ExecCfg().Codec, dbName, scName, tbName, lookupFlags)
+
+	// The returned object may contain types.T that need hydrating.
+	if objDesc != nil {
+		if err := p.maybeHydrateTypesInDescriptor(ctx, objDesc); err != nil {
+			return false, nil, err
+		}
+	}
+
+	return objDesc != nil, objDesc, err
+}
+
+func (p *planner) CommonLookupFlags(required bool) tree.CommonLookupFlags {
+	return tree.CommonLookupFlags{
+		Required:    required,
+		AvoidCached: p.avoidCachedDescriptors,
+	}
 }
 
 // ResolveType implements the TypeReferenceResolver interface.
@@ -149,7 +142,7 @@ func (p *planner) ResolveType(name *tree.UnresolvedObjectName) (*types.T, error)
 	}
 	// TODO (rohany): The ResolveAnyDescType argument doesn't do anything here
 	//  if we are looking for a type. This should be cleaned up.
-	desc, prefix, err := resolveExistingObjectImpl(p.EvalContext().Context, p, name, lookupFlags, ResolveAnyDescType)
+	desc, prefix, err := resolver.ResolveExistingObject(p.EvalContext().Context, p, name, lookupFlags, resolver.ResolveAnyDescType)
 	if err != nil {
 		return nil, err
 	}
@@ -240,209 +233,6 @@ func (p *planner) maybeHydrateTypesInDescriptor(
 	return nil
 }
 
-func resolveExistingObjectImpl(
-	ctx context.Context,
-	sc SchemaResolver,
-	un *tree.UnresolvedObjectName,
-	lookupFlags tree.ObjectLookupFlags,
-	requiredType ResolveRequiredType,
-) (res tree.NameResolutionResult, prefix tree.ObjectNamePrefix, err error) {
-	found, prefix, descI, err := tree.ResolveExisting(ctx, un, sc, lookupFlags, sc.CurrentDatabase(), sc.CurrentSearchPath())
-	if err != nil {
-		return nil, prefix, err
-	}
-	// Construct the resolved table name for use in error messages.
-	resolvedTn := tree.MakeTableNameFromPrefix(prefix, tree.Name(un.Object()))
-	if !found {
-		if lookupFlags.Required {
-			return nil, prefix, sqlbase.NewUndefinedObjectError(&resolvedTn, lookupFlags.DesiredObjectKind)
-		}
-		return nil, prefix, nil
-	}
-
-	obj := descI.(catalog.ObjectDescriptor)
-	switch lookupFlags.DesiredObjectKind {
-	case tree.TypeObject:
-		if obj.TypeDesc() == nil {
-			return nil, prefix, sqlbase.NewUndefinedTypeError(&resolvedTn)
-		}
-		return obj.TypeDesc(), prefix, nil
-	case tree.TableObject:
-		if obj.TableDesc() == nil {
-			return nil, prefix, sqlbase.NewUndefinedRelationError(&resolvedTn)
-		}
-		goodType := true
-		switch requiredType {
-		case ResolveRequireTableDesc:
-			goodType = obj.TableDesc().IsTable()
-		case ResolveRequireViewDesc:
-			goodType = obj.TableDesc().IsView()
-		case ResolveRequireTableOrViewDesc:
-			goodType = obj.TableDesc().IsTable() || obj.TableDesc().IsView()
-		case ResolveRequireSequenceDesc:
-			goodType = obj.TableDesc().IsSequence()
-		}
-		if !goodType {
-			return nil, prefix, sqlbase.NewWrongObjectTypeError(&resolvedTn, requiredTypeNames[requiredType])
-		}
-
-		// If the table does not have a primary key, return an error
-		// that the requested descriptor is invalid for use.
-		if !lookupFlags.AllowWithoutPrimaryKey &&
-			obj.TableDesc().IsTable() &&
-			!obj.TableDesc().HasPrimaryKey() {
-			return nil, prefix, errNoPrimaryKey
-		}
-
-		if lookupFlags.RequireMutable {
-			return descI.(*MutableTableDescriptor), prefix, nil
-		}
-
-		return descI.(*ImmutableTableDescriptor), prefix, nil
-	default:
-		return nil, prefix, errors.AssertionFailedf(
-			"unknown desired object kind %d", lookupFlags.DesiredObjectKind)
-	}
-}
-
-// runWithOptions sets the provided resolution flags for the
-// duration of the call of the passed argument fn.
-//
-// This is meant to be used like this (for example):
-//
-// var someVar T
-// var err error
-// p.runWithOptions(resolveFlags{skipCache: true}, func() {
-//    someVar, err = ResolveExistingTableObject(ctx, p, ...)
-// })
-// if err != nil { ... }
-// use(someVar)
-func (p *planner) runWithOptions(flags resolveFlags, fn func()) {
-	if flags.skipCache {
-		defer func(prev bool) { p.avoidCachedDescriptors = prev }(p.avoidCachedDescriptors)
-		p.avoidCachedDescriptors = true
-	}
-	fn()
-}
-
-type resolveFlags struct {
-	skipCache bool
-}
-
-func (p *planner) ResolveMutableTableDescriptor(
-	ctx context.Context, tn *TableName, required bool, requiredType ResolveRequiredType,
-) (table *MutableTableDescriptor, err error) {
-	return ResolveMutableExistingTableObject(ctx, p, tn, required, requiredType)
-}
-
-func (p *planner) ResolveUncachedTableDescriptor(
-	ctx context.Context, tn *TableName, required bool, requiredType ResolveRequiredType,
-) (table *ImmutableTableDescriptor, err error) {
-	p.runWithOptions(resolveFlags{skipCache: true}, func() {
-		lookupFlags := tree.ObjectLookupFlags{CommonLookupFlags: tree.CommonLookupFlags{Required: required}}
-		table, err = ResolveExistingTableObject(ctx, p, tn, lookupFlags, requiredType)
-	})
-	return table, err
-}
-
-// ResolveTargetObject determines a valid target path for an object
-// that may not exist yet. It returns the descriptor for the database
-// where the target object lives. It also returns the resolved name
-// prefix for the input object.
-func ResolveTargetObject(
-	ctx context.Context, sc SchemaResolver, un *tree.UnresolvedObjectName,
-) (*DatabaseDescriptor, tree.ObjectNamePrefix, error) {
-	found, prefix, descI, err := tree.ResolveTarget(ctx, un, sc, sc.CurrentDatabase(), sc.CurrentSearchPath())
-	if err != nil {
-		return nil, prefix, err
-	}
-	if !found {
-		if !un.HasExplicitSchema() && !un.HasExplicitCatalog() {
-			return nil, prefix, pgerror.New(pgcode.InvalidName, "no database specified")
-		}
-		err = pgerror.Newf(pgcode.InvalidSchemaName,
-			"cannot create %q because the target database or schema does not exist",
-			tree.ErrString(un))
-		err = errors.WithHint(err, "verify that the current database and search_path are valid and/or the target database exists")
-		return nil, prefix, err
-	}
-	if prefix.Schema() != tree.PublicSchema {
-		return nil, prefix, pgerror.Newf(pgcode.InvalidName,
-			"schema cannot be modified: %q", tree.ErrString(&prefix))
-	}
-	return descI.(*DatabaseDescriptor), prefix, nil
-}
-
-func (p *planner) ResolveUncachedDatabase(
-	ctx context.Context, un *tree.UnresolvedObjectName,
-) (res *UncachedDatabaseDescriptor, namePrefix tree.ObjectNamePrefix, err error) {
-	p.runWithOptions(resolveFlags{skipCache: true}, func() {
-		res, namePrefix, err = ResolveTargetObject(ctx, p, un)
-	})
-	return res, namePrefix, err
-}
-
-// ResolveRequiredType can be passed to the ResolveExistingTableObject function to
-// require the returned descriptor to be of a specific type.
-type ResolveRequiredType int
-
-// ResolveRequiredType options have descriptive names.
-const (
-	ResolveAnyDescType ResolveRequiredType = iota
-	ResolveRequireTableDesc
-	ResolveRequireViewDesc
-	ResolveRequireTableOrViewDesc
-	ResolveRequireSequenceDesc
-)
-
-var requiredTypeNames = [...]string{
-	ResolveRequireTableDesc:       "table",
-	ResolveRequireViewDesc:        "view",
-	ResolveRequireTableOrViewDesc: "table or view",
-	ResolveRequireSequenceDesc:    "sequence",
-}
-
-// LookupSchema implements the tree.ObjectNameTargetResolver interface.
-func (p *planner) LookupSchema(
-	ctx context.Context, dbName, scName string,
-) (found bool, scMeta tree.SchemaMeta, err error) {
-	sc := p.LogicalSchemaAccessor()
-	dbDesc, err := sc.GetDatabaseDesc(ctx, p.txn, p.ExecCfg().Codec, dbName, p.CommonLookupFlags(false /*required*/))
-	if err != nil || dbDesc == nil {
-		return false, nil, err
-	}
-	found, _, err = sc.IsValidSchema(ctx, p.txn, p.ExecCfg().Codec, dbDesc.ID, scName)
-	if err != nil {
-		return false, nil, err
-	}
-	return found, dbDesc, nil
-}
-
-// LookupObject implements the tree.ObjectNameExistingResolver interface.
-func (p *planner) LookupObject(
-	ctx context.Context, lookupFlags tree.ObjectLookupFlags, dbName, scName, tbName string,
-) (found bool, objMeta tree.NameResolutionResult, err error) {
-	sc := p.LogicalSchemaAccessor()
-	lookupFlags.CommonLookupFlags = p.CommonLookupFlags(false /* required */)
-	objDesc, err := sc.GetObjectDesc(ctx, p.txn, p.ExecCfg().Settings, p.ExecCfg().Codec, dbName, scName, tbName, lookupFlags)
-
-	// The returned object may contain types.T that need hydrating.
-	if objDesc != nil {
-		if err := p.maybeHydrateTypesInDescriptor(ctx, objDesc); err != nil {
-			return false, nil, err
-		}
-	}
-
-	return objDesc != nil, objDesc, err
-}
-
-func (p *planner) CommonLookupFlags(required bool) tree.CommonLookupFlags {
-	return tree.CommonLookupFlags{
-		Required:    required,
-		AvoidCached: p.avoidCachedDescriptors,
-	}
-}
-
 func (p *planner) ObjectLookupFlags(required, requireMutable bool) tree.ObjectLookupFlags {
 	return tree.ObjectLookupFlags{
 		CommonLookupFlags: p.CommonLookupFlags(required),
@@ -485,12 +275,13 @@ func getDescriptorsFromTargetList(
 		if err != nil {
 			return nil, err
 		}
+
 		for i := range objectNames {
 			// We set required to false here, because there could be type names in
 			// the returned set of names, so we don't want to error out if we
 			// couldn't resolve a name into a table.
-			descriptor, err := ResolveMutableExistingTableObject(
-				ctx, p, &objectNames[i], false /* required */, ResolveAnyDescType)
+			descriptor, err := resolver.ResolveMutableExistingTableObject(ctx, p,
+				&objectNames[i], false /* required */, resolver.ResolveAnyDescType)
 			if err != nil {
 				return nil, err
 			}
@@ -539,7 +330,7 @@ func (p *planner) getQualifiedTableName(
 func findTableContainingIndex(
 	ctx context.Context,
 	txn *kv.Txn,
-	sc SchemaResolver,
+	sc resolver.SchemaResolver,
 	codec keys.SQLCodec,
 	dbName, scName string,
 	idxName tree.UnrestrictedName,
@@ -560,7 +351,7 @@ func findTableContainingIndex(
 	result = nil
 	for i := range tns {
 		tn := &tns[i]
-		tableDesc, err := ResolveMutableExistingTableObject(ctx, sc, tn, false /*required*/, ResolveAnyDescType)
+		tableDesc, err := resolver.ResolveMutableExistingTableObject(ctx, sc, tn, false /*required*/, resolver.ResolveAnyDescType)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -610,7 +401,7 @@ func expandMutableIndexName(
 func expandIndexName(
 	ctx context.Context,
 	txn *kv.Txn,
-	sc SchemaResolver,
+	sc resolver.SchemaResolver,
 	codec keys.SQLCodec,
 	index *tree.TableIndexName,
 	requireTable bool,
@@ -618,7 +409,7 @@ func expandIndexName(
 	tn = &index.Table
 	if tn.Table() != "" {
 		// The index and its table prefix must exist already. Resolve the table.
-		desc, err = ResolveMutableExistingTableObject(ctx, sc, tn, requireTable, ResolveRequireTableDesc)
+		desc, err = resolver.ResolveMutableExistingTableObject(ctx, sc, tn, requireTable, resolver.ResolveRequireTableDesc)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -702,12 +493,12 @@ func expandTableGlob(
 // of CREATE TABLE with a foreign key self-reference: the target of
 // the FK definition is a table that does not exist yet.
 type fkSelfResolver struct {
-	SchemaResolver
+	resolver.SchemaResolver
 	newTableName *tree.TableName
 	newTableDesc *sqlbase.TableDescriptor
 }
 
-var _ SchemaResolver = &fkSelfResolver{}
+var _ resolver.SchemaResolver = &fkSelfResolver{}
 
 // LookupObject implements the tree.ObjectNameExistingResolver interface.
 func (r *fkSelfResolver) LookupObject(
@@ -859,10 +650,10 @@ func (p *planner) ResolveMutableTableDescriptorEx(
 	ctx context.Context,
 	name *tree.UnresolvedObjectName,
 	required bool,
-	requiredType ResolveRequiredType,
+	requiredType resolver.ResolveRequiredType,
 ) (*MutableTableDescriptor, error) {
 	tn := name.ToTableName()
-	table, err := ResolveMutableExistingTableObject(ctx, p, &tn, required, requiredType)
+	table, err := resolver.ResolveMutableExistingTableObject(ctx, p, &tn, required, requiredType)
 	if err != nil {
 		return nil, err
 	}
@@ -877,14 +668,14 @@ func (p *planner) ResolveMutableTableDescriptorExAllowNoPrimaryKey(
 	ctx context.Context,
 	name *tree.UnresolvedObjectName,
 	required bool,
-	requiredType ResolveRequiredType,
+	requiredType resolver.ResolveRequiredType,
 ) (*MutableTableDescriptor, error) {
 	lookupFlags := tree.ObjectLookupFlags{
 		CommonLookupFlags:      tree.CommonLookupFlags{Required: required},
 		RequireMutable:         true,
 		AllowWithoutPrimaryKey: true,
 	}
-	desc, prefix, err := resolveExistingObjectImpl(ctx, p, name, lookupFlags, requiredType)
+	desc, prefix, err := resolver.ResolveExistingObject(ctx, p, name, lookupFlags, requiredType)
 	if err != nil || desc == nil {
 		return nil, err
 	}
@@ -898,7 +689,7 @@ func (p *planner) ResolveUncachedTableDescriptorEx(
 	ctx context.Context,
 	name *tree.UnresolvedObjectName,
 	required bool,
-	requiredType ResolveRequiredType,
+	requiredType resolver.ResolveRequiredType,
 ) (table *ImmutableTableDescriptor, err error) {
 	p.runWithOptions(resolveFlags{skipCache: true}, func() {
 		table, err = p.ResolveExistingObjectEx(ctx, name, required, requiredType)
@@ -911,10 +702,10 @@ func (p *planner) ResolveExistingObjectEx(
 	ctx context.Context,
 	name *tree.UnresolvedObjectName,
 	required bool,
-	requiredType ResolveRequiredType,
+	requiredType resolver.ResolveRequiredType,
 ) (res *ImmutableTableDescriptor, err error) {
 	lookupFlags := tree.ObjectLookupFlags{CommonLookupFlags: tree.CommonLookupFlags{Required: required}}
-	desc, prefix, err := resolveExistingObjectImpl(ctx, p, name, lookupFlags, requiredType)
+	desc, prefix, err := resolver.ResolveExistingObject(ctx, p, name, lookupFlags, requiredType)
 	if err != nil || desc == nil {
 		return nil, err
 	}
